Moved range building to  separate class
